fix: address blocking and non-blocking review findings in doctor command
Blocking fixes:
- Fix data loss in fixLayout catch block: only delete tmpDbPath if the original
  flat file still exists; otherwise leave it as the sole surviving copy and
  surface its location in the error message for manual recovery
- Fix stale resolvedDbPath after fixLayout: re-call resolveDbPath(absRaw) after
  a successful layout fix so subsequent fixSchema/fixFtsIndex use the new path

Non-blocking fixes:
- Extract all check/fix functions from doctor.ts into doctor-checks.ts to keep
  each file under 500 lines (Fix 3)
- Add explicit comment in fixLayout JSDoc explaining the intentional gitignore
  side effect (Fix 4)
- Rename test description "warns" to "fails" for the flat-file gitignore case
  that correctly calls fail() (Fix 5)
- Derive WAL/SHM file names from --db path in checkWal/fixWal rather than
  hardcoding .engram-wal/.engram-shm (Fix 6)
- Add FORMAT_VERSION guard in fixSchema with a warning if version has advanced
  beyond the highest known migration (Fix 7)
- Restore process.exit in finally blocks in all tests (Fix 8)
- Add afterEach cleanup of temp directories in all test suites (Fix 9)
